AddEditSpot screen viewmodel done

+++ data/src/main/java/com/harbdm/urbanexplorer/data/local/db/entity/SpotEntity.kt
@@ -18,7 +18,7 @@ data class SpotEntity(
     @ColumnInfo(name = "spot_rating")
     val spotRating: Int,
     @ColumnInfo(name = "time_stamp")
-    val timeStamp: Int,
+    val timeStamp: Long,
     @PrimaryKey(autoGenerate = true)
     val id: Long = 0
 )
